rtl8723au: Upgrade to driver version v4.1.6_7336.20130426
This update replaces the entire source and should be tested. If it fails,
report it and do a "git checkout 4682dce" to return to the previous
version.

+++ include/usb_osintf.h
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Copyright(c) 2007 - 2011 Realtek Corporation. All rights reserved.
- *
+ *                                        
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of version 2 of the GNU General Public License as
  * published by the Free Software Foundation.
@@ -25,7 +25,7 @@
 #include <drv_types.h>
 #include <usb_vendor_req.h>
 
-#define USBD_HALTED(Status) ((u32)(Status) >> 30 == 3)
+#define USBD_HALTED(Status) ((ULONG)(Status) >> 30 == 3)
 
 
 //uint usb_dvobj_init(_adapter * adapter);
@@ -35,3 +35,4 @@ u8 usbvendorrequest(struct dvobj_priv *pdvobjpriv, RT_USB_BREQUEST brequest, RT_
 
 
 #endif
+
